Add back drawHighlighted

+++ src/imagemgr.cpp
@@ -341,7 +341,7 @@ void ImageMgr::fixupIntro(Image *im, int prescale) {
      * draw "Lord British" signature
      * ----------------------------- */
     color = RGBA{0, 255, 255, 255};  // cyan for EGA
-    unsigned int blue[16] = {255, 250, 226, 226, 210, 194, 161, 161,
+    uint8_t blue[16] = {255, 250, 226, 226, 210, 194, 161, 161,
                     129,  97,  97,  64,  64,  32,  32,   0};
     i = 0;
     while (sigData[i] != 0) {
@@ -352,7 +352,7 @@ void ImageMgr::fixupIntro(Image *im, int prescale) {
         if (settings.videoType) // Not EGA
         {
             // yellow gradient
-            color = RGBA{255, (unsigned int)(y == 1 ? 250 : 255), blue[y], 255};
+            color = RGBA{255, (uint8_t)(y == 1 ? 250 : 255), blue[y], 255};
         }
 
         im->fillRect(x * prescale, y * prescale,
